fix(map): rename Pass-4 support helper so dart test does not load it
fill_lakes_pass4_for_test.dart matched *_test.dart discovery and failed
CI with Missing definition of main; rename to *_support.dart.
