+++ packages/colonizethis_map/test/tile_map_pass4_lakes_test.dart
@@ -2,7 +2,7 @@ import 'package:colonizethis_test/test.dart';
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_map/src/tile_map_directions.dart';
 
-import 'support/fill_lakes_pass4_for_test.dart';
+import 'support/fill_lakes_pass4_support.dart';
 import 'support/tile_map_gen_fixtures.dart';
 
 void main() {
